Add moon gravity, organize plugin files
