+++ src/main/java/com/dinnerbone/bukkit/moon/command/MoonCommandExec.java
@@ -1,5 +1,5 @@
 
-package com.dinnerbone.bukkit.moon;
+package com.dinnerbone.bukkit.moon.command;
 
 import com.dinnerbone.bukkit.moon.BukkitMoon;
 import org.bukkit.ChatColor;
@@ -8,6 +8,8 @@ import org.bukkit.command.Command;
 import org.bukkit.command.CommandExecutor;
 import org.bukkit.command.CommandSender;
 import org.bukkit.entity.Player;
+import org.bukkit.potion.PotionEffect;
+import org.bukkit.potion.PotionEffectType;
 
 import java.util.HashMap;
 import java.util.Map;
@@ -26,6 +28,8 @@ public class MoonCommandExec implements CommandExecutor {
 
             worlds.put(player, player.getWorld());
             player.teleport(BukkitMoon.getMoon().getSpawnLocation());
+            //gravity
+            player.addPotionEffect(new PotionEffect(PotionEffectType.JUMP, Integer.MAX_VALUE, Double.valueOf(2.66).intValue(), false));
         } else {
             sender.sendMessage("I don't know who you are!");
         }
