+++ src/main/java/com/dinnerbone/bukkit/moon/terrain/MoonChunkGenerator.java
@@ -1,9 +1,10 @@
 
-package com.dinnerbone.bukkit.moon;
+package com.dinnerbone.bukkit.moon.terrain;
 
 import java.util.Arrays;
 import java.util.List;
 import java.util.Random;
+
 import org.bukkit.Location;
 import org.bukkit.generator.BlockPopulator;
 import org.bukkit.generator.ChunkGenerator;
@@ -30,7 +31,7 @@ public class MoonChunkGenerator extends ChunkGenerator {
         result *= variance;
         return NoiseGenerator.floor(result);
     }
-    
+
     public byte[] generate(World world, Random random, int cx, int cz) {
         byte[] result = new byte[32768];
